\name{construct.v}
\alias{construct.v}
\title{Split Names of Vectors into N-vectors
This function is used to organize how variables are spread
across the list of formulas}
\usage{
  construct.v(constraints, ndim)
}
\arguments{
  \item{constraints}{a constraints object}

  \item{ndim}{}
}
\value{
  a list of character-vectors
}
\description{
  Split Names of Vectors into N-vectors This function is
  used to organize how variables are spread across the list
  of formulas
}
\author{
  Matt Owen and Olivia Lau and Kosuke Imai
}

