\name{describe.blogit}
\alias{describe.blogit}
\title{Describe the Citation of the \code{blogit} Model}
\usage{
  \method{describe}{blogit}(...)
}
\arguments{
  \item{...}{dummy parameters}
}
\value{
  a \code{description} object used to cite the
  \code{blogit} model
}
\description{
  Describe the Citation of the \code{blogit} Model
}
\author{
  Matt Owen, Olivia Lau, and Kosuke Imai
}

