\name{describe.bprobit}
\alias{describe.bprobit}
\title{Describe the Citation of the \code{bprobit} Model}
\usage{
  \method{describe}{bprobit}(...)
}
\arguments{
  \item{...}{dummy parameters}
}
\value{
  a \code{description} object used to cite the
  \code{bprobit} model
}
\description{
  Describe the Citation of the \code{bprobit} Model
}
\author{
  Matt Owen, Olivia Lau, and Kosuke Imai
}

