\name{describe.mlogit}
\alias{describe.mlogit}
\title{Provide Citation Information for the ``blogit'' Model}
\usage{
  \method{describe}{mlogit}(...)
}
\arguments{
  \item{...}{dummy parameters}
}
\value{
  a list
}
\description{
  Provide Citation Information for the ``blogit'' Model
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

