\name{describe.mprobit}
\alias{describe.mprobit}
\title{Describe the Multinomial Probit Model}
\usage{
  describe.mprobit(...)
}
\arguments{
  \item{...}{dummy parameters}
}
\value{
  a list
}
\description{
  Describe the Multinomial Probit Model
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

