\name{describe.ologit}
\alias{describe.ologit}
\title{describe a \code{ologit} model to zelig}
\usage{
  describe.ologit(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list to be processed by \code{as.description}
}
\description{
  describe a \code{ologit} model to zelig
}
\author{
  matt owen \email{mowen@iq.harvard.edu}
}

