\name{ev.mlogit}
\alias{ev.mlogit}
\title{Simulate Expected Value for Multinomial Logit}
\usage{
  ev.mlogit(fitted, constraints, all.coef, x, ndim, cnames)
}
\arguments{
  \item{fitted}{a fitted model object}

  \item{constraints}{a constraints object}

  \item{all.coef}{all the coeficients}

  \item{x}{a setx object}

  \item{ndim}{an integer specifying the number of
  dimensions}

  \item{cnames}{a character-vector specifying the names of
  the columns}
}
\value{
  a matrix of simulated values
}
\description{
  Simulate Expected Value for Multinomial Logit
}
\author{
  Matt Owen and Olivia Lau and Kosuke Imai
  \email{mowen@iq.harvard.edu}
}

