\name{param.bprobit}
\alias{param.bprobit}
\title{Param Method for the 'bprobit' Zelig Model}
\usage{
  param.bprobit(obj, num = 1000, ...)
}
\arguments{
  \item{obj}{a 'zelig' object}

  \item{num}{an integer specifying the number of
  simulations to sample}

  \item{...}{ignored parameters}
}
\value{
  a list to be cast as a 'parameters' object
}
\description{
  Param Method for the 'bprobit' Zelig Model
}
\note{
  This method is used by the 'bprobit' Zelig model
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

