\name{plot.sim.ologit}
\alias{plot.sim.ologit}
\title{plot.sim.ologit}
\usage{
  plot.sim.ologit(x, ..., alt.color = NULL)
}
\arguments{
  \item{x}{a \code{sim.ologit} object}

  \item{...}{ignored parameters}

  \item{alt.color}{list of colors to print with}
}
\value{
  function is used for its side-effects
}
\description{
  plot.sim.ologit
}

