\name{plot.sim.oprobit}
\alias{plot.sim.oprobit}
\title{plot.sim.oprobit}
\usage{
  plot.sim.oprobit(x, ..., alt.color = NULL)
}
\arguments{
  \item{x}{a \code{sim.oprobit} object}

  \item{...}{ignored parameters}

  \item{alt.color}{list of colors to print with}
}
\value{
  function is used for its side-effects
}
\description{
  plot.sim.oprobit
}

