\name{pv.mlogit}
\alias{pv.mlogit}
\title{Simulate Predicted Values}
\usage{
  pv.mlogit(fitted, ev, ynames)
}
\arguments{
  \item{fitted}{a fitted model object}

  \item{ynames}{???}
}
\value{
  a vector of simulated values
}
\description{
  Simulate Predicted Values
}
\author{
  Matt Owen and Olivia Lau and Kosuke Imai
  \email{mowen@iq.harvard.edu}
}

