\name{summarize.ologit}
\alias{summarize.ologit}
\title{summarize.ologit
summarizes simulations of quantities of interest}
\usage{
  summarize.ologit(qi)
}
\arguments{
  \item{qi}{a quantity of interest object}
}
\description{
  summarize.ologit summarizes simulations of quantities of
  interest
}

