\name{summarize.oprobit}
\alias{summarize.oprobit}
\title{summarize.ologit
summarizes simulations of quantities of interest}
\usage{
  summarize.oprobit(qi)
}
\arguments{
  \item{qi}{a quantity of interest object}
}
\description{
  summarize.ologit summarizes simulations of quantities of
  interest
}

