\name{zelig2bprobit}
\alias{zelig2bprobit}
\title{Interface between \code{bprobit} model and Zelig
This function is exclusively for use by the \code{zelig} function}
\usage{
  zelig2bprobit(formula, ..., data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{ignored parameters}

  \item{data}{a data.frame}
}
\value{
  a list to be coerced into a zelig.call object
}
\description{
  Interface between \code{bprobit} model and Zelig This
  function is exclusively for use by the \code{zelig}
  function
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

