\name{zelig2mlogit}
\alias{zelig2mlogit}
\title{interface between the Zelig model mlogit and the pre-existing function}
\usage{
  zelig2mlogit(formula, ..., data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{ignored parameters}

  \item{data}{a data.frame}
}
\value{
  a list specifying '.function'
}
\description{
  interface between the Zelig model mlogit and the
  pre-existing function
}

