\docType{package}
\name{ZeligMultinomial-package.R}
\alias{ZeligMultinomial}
\alias{ZeligMultinomial-package.R}
\title{Multinomial Logit and Probit Zelig Models}
\description{
  \tabular{ll}{ Package: \tab ZeligMultinomial\cr Version:
  \tab 0.5-4\cr Date: \tab 2011-05-04\cr Depends: Zelig (>=
  4.0-6), MNP, VGAM\cr License: \tab GPL version 2 or
  newer\cr URL: \tab http://gking.harvard.edu/zelig\cr }
}
\details{
  Add-on pack for Zelig, containing a multinomial models:
  \tabular{ll}{ mlogit: \tab Multinomial Logistic Model,
  implemented by VGAM \cr mprobit: \tab Multinomial Probit
  Model, implemented by MNP \cr }
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}, Kosuke Imai,
  Olivia Lau, and Gary King

  Maintainer: Matt Owen \email{mowen@iq.harvard.edu}
}
\keyword{package}

