\name{ZeligMultivariate-package}
\title{Bivariate Zelig Models...}
\description{Bivariate Zelig Models}
\details{\tabular{ll}{
Package: \tab ZeligMultivariate\cr
Version: \tab 1.0-4\cr
Date: \tab 2012-01-12\cr
Depends: \tab Zelig (>= 4.0-6), VGAM (>= 0.8-4)\cr
Suggests: mvtnorm, MCMCpack\cr
License: \tab GPL version 2 or newer\cr
URL: \tab http://gking.harvard.edu/zelig\cr
}

Add-on pack for Zelig, containing a bivariate probit and logit regession}
\alias{ZeligMultivariate}
\docType{package}
\author{Matt Owen \email{mowen@iq.harvard.edu}, Kosuke Imai, Olivia Lau,
and Gary King 

Maintainer: Matt Owen \email{mowen@iq.harvard.edu}}
\keyword{package}
