\docType{package}
\name{ZeligOrdinal-package}
\alias{ZeligOrdinal}
\alias{ZeligOrdinal-package}
\title{Ordinal Binomial Models in Zelig}
\description{
  \tabular{ll}{ Package: \tab ZeligOrdinal\cr Version: \tab
  1.0-4\cr Date: \tab 2011-05-04\cr Depends: Zelig (>=
  4.0-6), MASS\cr License: \tab GPL version 2 or newer\cr
  URL: \tab http://gking.harvard.edu/zelig\cr }
}
\details{
  Add-on pack for Zelig, containing ordinal models:
  \tabular{ll}{ ologit: \tab Ordinal Logit Regression \cr
  oprobit: \tab Ordinal Probit Regression \cr }
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}, Kosuke Imai,
  Olivia Lau, and Gary King

  Maintainer: Matt Owen \email{mowen@iq.harvard.edu}
}
\keyword{package}

