\name{describe.blogit}
\alias{describe.blogit}
\title{Describe the Citation of the blogit Model...}
\usage{describe.blogit(...)}
\description{Describe the Citation of the \code{blogit} Model}
\value{a \code{description} object used to cite the \code{blogit} model}
\author{Matt Owen, Olivia Lau, and Kosuke Imai}
\arguments{\item{...}{dummy parameters}}
