\name{describe.bprobit}
\alias{describe.bprobit}
\title{Describe the Citation of the bprobit Model...}
\usage{describe.bprobit(...)}
\description{Describe the Citation of the \code{bprobit} Model}
\value{a \code{description} object used to cite the \code{bprobit} model}
\author{Matt Owen, Olivia Lau, and Kosuke Imai}
\arguments{\item{...}{dummy parameters}}
