\name{zelig2blogit}
\alias{zelig2blogit}
\title{Interface between \code{blogit} model and Zelig
This function is exclusively for use by the \code{zelig} function}
\usage{
  zelig2blogit(formula, ..., data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{ignored parameters}

  \item{data}{a data.frame}
}
\value{
  a list to be coerced into a zelig.call object
}
\description{
  Interface between \code{blogit} model and Zelig This
  function is exclusively for use by the \code{zelig}
  function
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

