\name{zelig2blogit}
\alias{zelig2blogit}
\title{Interface between blogit model and Zelig...}
\usage{zelig2blogit(formula, ..., data)}
\description{Interface between \code{blogit} model and Zelig
This function is exclusively for use by the \code{zelig} function}
\value{a list to be coerced into a zelig.call object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{formula}{a formula}
\item{...}{ignored parameters}
\item{data}{a data.frame}}
