\name{zelig2bprobit}
\alias{zelig2bprobit}
\title{Interface between bprobit model and Zelig...}
\usage{zelig2bprobit(formula, ..., data)}
\description{Interface between \code{bprobit} model and Zelig
This function is exclusively for use by the \code{zelig} function}
\value{a list to be coerced into a zelig.call object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{formula}{a formula}
\item{...}{ignored parameters}
\item{data}{a data.frame}}
